refactor: slice pattern 적용
